requirements.txt
absl-py==0.8.1
appnope==0.1.0
argcomplete==1.10.0
astor==0.8.0
attrs==19.1.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
blis==0.4.1
boltons==19.1.0
boto==2.49.0
boto3==1.10.34
botocore==1.13.34
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
configparser==4.0.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
Django==2.1.15
django-cors-headers==3.1.1
djangorestframework==3.10.3
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
gast==0.2.2
gensim==3.8.1
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
IMAPClient==2.1.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lenskit==0.7.0
llvmlite==0.30.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
mypy-extensions==0.4.3
mysql==0.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.6
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numba==0.46.0
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
pdfminer.six==20181108
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Protego==0.1.15
protobuf==3.10.0
ptyprocess==0.6.0
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.4
PyDispatcher==2.0.5
Pygments==2.4.2
pygooglechart==0.4.0
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.2
PyPDF2==1.26.0
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.0
python-pptx==0.6.18
python-sql==1.0.0
pytz==2019.2
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.5.5
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
Scrapy==1.8.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sklearn==0.0
smart-open==1.9.0
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.2.2
SpeechRecognition==3.8.1
sqlparse==0.3.0
srsly==0.2.0
stemming==1.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textract==1.6.3
thinc==7.3.1
tika==1.22
tornado==6.0.3
tqdm==4.37.0
traitlets==4.3.2
tweepy==3.8.0
Twisted==19.10.0
tzlocal==1.5.1
urllib3==1.25.6
w3lib==1.21.0
wasabi==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.6
zipp==0.6.0
zope.interface==4.7.1
